+++ celescope/tools/cellranger3/get_plot_elements.py
@@ -114,9 +114,6 @@ def segment_log_plot_by_length(y_data, x_start, x_end):
     this_segment_len = 0.0
     segment_idx = [x_start]
 
-    np.seterr(divide = 'ignore')
-    np.seterr(invalid='ignore')
-
     for i in range(x_start, x_end):
         last_i = max(x_start, i-1)
         dx = (np.log(i) - np.log(last_i)) / log_max_x