+++ celescope/trust_vdj/__init__.py
@@ -1,8 +1,0 @@
-__STEPS__ = [
-    'sample',
-    'barcode',
-    'cutadapt',
-    'convert',
-    'assemble',
-    'res_sum']
-__ASSAY__ = 'trust_vdj'