+++ tests/unittests/__init__.py
@@ -1,0 +1,1 @@
+TEST_DIR_ROOT = '/SGRNJ03/randd/user/zhouyiqi/multi_tests/'