+++ docs/CHANGELOG.md
@@ -1,9 +1,40 @@
-# Change Log
+## [unreleased] - 2021-08-17
+ ### Added
 
-## [unreleased] - 2021-06-09
+ ### Changed
+
+ ### Fixed
+
+ ### Removed
+
+## [1.4.0b0] - 2021-08-17
+ ### Added
+
+ - Add `min_consensus_read` argument to `celescope.tools.consensus`. If 
+    1. the percentage of the most common residue type > threshold;
+    2. most common residue reads >= min_consensus_read;
+    
+then we will add that residue type, otherwise an ambiguous character will be added.
+
+- Use otsu method to calculate `min_support_read` for `capture_virus`.
+
+ ### Changed
+
+ ### Fixed
+
+ - Fix an error when there is no `Undetermined` cells in `celescope tag count_tag`.
+
+ ### Removed
+
+## [1.3.2] - 2021-07-09
 ### Added
 
+- Add `dynascope` assay.
+- Add `celescope tag split_tag`. 
+
 ### Changed
+- Change fastq file pattern of mapfile: Remove * before library_id.
+- `celescope.tools.count_capture_virus`: Change `min_support_read` from 1 to 2.
 
 ### Fixed
 - `celescope.tools.count` will report an error when there are multiple gtf or refFlat file under `genomeDir`.